+++ module/These/src/These/Notification/ResultatTheseAdmisNotification.php
@@ -35,7 +35,8 @@ class ResultatTheseAdmisNotification extends Notification
      */
     public function prepare()
     {
-        $to = $this->these->getDoctorant()->getEmail();
+        $individu = $this->these->getDoctorant()->getIndividu();
+        $to = $individu->getEmailContact() ?: $individu->getEmailPro() ?: $individu->getEmailUtilisateur();
 
         $this->emailDoctorantAbsent = false;
         if (! $to) {